refactor: create separate svelte components for GitHub issue and developers feedback Signed-off-by: Sonia Sandler <ssandler@redhat.com>
